Add positional operator for $inc. Also, fix an issue with multiple edits and positional operators.
The problem was that the document would change after the first edit command.
This would result in the spec not matching on the second edit command and
thus the second edit not succeeding.

So we now store the subdocument in the update method. If the subdocument is
found, we can simply reuse it.