Last fixup for H405 error
Required a lot of work for, in my opinion, a very superfluous coding
standard. To beat it I changed the docstring to be a single line
comment. It made the comment less readable, but it got me past H405.

The bright side, I was able to recreate the issue on my machine. It
required making a change to pyflakes/checker.py. For anyone experiencing
the same problem, here is the fix I used:

In pyflakes 1.2.3, file pyflakes/checker.py, change line 854

from:

        STARRED = NAMECONSTANT = handleChildren
to:
        STARRED = NAMECONSTANT = CONSTANT = handleChildren