Minor modifications to support disk-based stores
Previously when performing update operations the document dictionary
would be updated in-place (and rolled back if need be, say on conflict).
This approach makes it difficult to implement a store where the
documents are not in-memory, for example, stored on disk.  In this case,
the store just returns copies of the real dictionaries and any mutations
are not reflected in the real document on disk.  This commit gets around
this by saving the mutated dictionary back to the collection once the
code is finished updating it.